añadida la verificación en edit y forms

+++ src/edit.py
@@ -2,10 +2,27 @@ import string
 from flask import Flask, flash, render_template, request, redirect, url_for, flash, session
 from flask_mysqldb import MySQL
 
-from __main__ import app, mysql, e_id
+from __main__ import app, mysql
+
+def verificar(id):
+    if 'loggedin' not in session:
+        return False
+    if session['type'] == "encuestado":
+        return False
+    cur = mysql.connection.cursor()
+    cur.execute("SELECT * FROM encuesta WHERE id_encuesta = %s AND id_encuestador = %s;",(id, session['id'],))
+    validar = cur.fetchall()
+    if validar == ():
+        return False
+    return True
 
 @app.route('/edit_respuesta/<id>/<id_pregunta>/<id_respuesta>', methods=['GET','POST'])
 def edit_respuesta(id,id_pregunta,id_respuesta):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+    # editar respuesta
     if request.method == 'POST':
         if(request.form['respuesta']!=""):
             cur = mysql.connection.cursor()
@@ -20,6 +37,11 @@ def edit_respuesta(id,id_pregunta,id_respuesta):
 
 @app.route('/edit_pregunta/<id>/<id_pregunta>', methods=['GET','POST'])
 def edit_pregunta(id,id_pregunta):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+    #editar pregunta
     if request.method == 'POST':
         if(request.form['pregunta']!=""):
             cur = mysql.connection.cursor()
@@ -34,6 +56,11 @@ def edit_pregunta(id,id_pregunta):
 
 @app.route('/edit_titulo/<id>', methods=['GET','POST'])
 def edit_titulo(id):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+    # editar titulo
     if request.method == 'POST':
         if(request.form['titulo']!=""):
             cur = mysql.connection.cursor()
@@ -49,9 +76,14 @@ def edit_titulo(id):
 
 @app.route('/new_alternativa/<id>/<id_pregunta>', methods=['GET','POST'])
 def new_alternativa(id,id_pregunta):
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+
+    #agregar alternativa
     if request.method == 'POST':
         if(request.form['alternativa']!=""):
-            print(request.form['alternativa'])
             cur = mysql.connection.cursor()
             cur.execute("INSERT INTO respuesta (id_pregunta, texto_respuesta) VALUES (%s , %s);", (id_pregunta ,request.form['alternativa'],))
             mysql.connection.commit()
@@ -63,9 +95,13 @@ def new_alternativa(id,id_pregunta):
 
 @app.route('/delete_alternativa/<id>/<id_alternativa>')
 def delete_alternativa(id,id_alternativa):
-    cur = mysql.connection.cursor()
-    
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
+
     #borrar la respuesta
+    cur = mysql.connection.cursor()
     cur.execute("DELETE FROM respuesta WHERE id_respuesta = %s;", (id_alternativa,))
     mysql.connection.commit()
     flash("Respuesta eliminada.")
@@ -73,9 +109,13 @@ def delete_alternativa(id,id_alternativa):
 
 @app.route('/delete_pregunta/<id>/<id_pregunta>')
 def delete_pregunta(id,id_pregunta):
-    cur = mysql.connection.cursor()
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
 
     #borrar todas las respuestas vinculados a la pregunta
+    cur = mysql.connection.cursor()
     cur.execute("DELETE FROM respuesta WHERE id_pregunta = %s;", (id_pregunta,))
     mysql.connection.commit()
     
@@ -87,7 +127,7 @@ def delete_pregunta(id,id_pregunta):
 
 @app.route('/edit/<id>', methods=['GET','POST'])
 def edit(id):
-
+    cur = mysql.connection.cursor()
     #Prohibido acceder a personas que no están registradas.
     if 'loggedin' not in session:
         return(render_template("403.html"))
@@ -97,7 +137,13 @@ def edit(id):
     #Los encuestados tampoco deberían editar las encuestas.
     if(tiposesion == "encuestado"):
         return(render_template("403.html"))
+    
+    # verificar
+    if not verificar(id):
+        flash("Usted no puede editar esta encuesta.")
+        return redirect(url_for("forms"))
 
+    #edit
     if request.method == 'POST':
         if(request.form['pregunta'] != ""):
             tipo_pregunta = -1;
@@ -117,8 +163,6 @@ def edit(id):
 
     #Visualización
     else:
-        cur = mysql.connection.cursor()
-
         #Encontrar el nombre de la encuesta en db
         cur.execute("SELECT titulo FROM encuesta WHERE id_encuesta = " + id)
         nombre_encuesta = cur.fetchall()
